First cut of new physics-related methods on world elements.
World elements now have the following additional methods:

 - applyImpluse(velocity, forcePoint)
 - setPersistentForce(force)
 - clearPersistentForce()
 - setVelocity(velocity)

However, the provisioning of this API is pretty ad-hoc and could be woven into the rest of the application better. Because we have new entities and attributes in the scene graph, keeping the physics engine at arms length may no longer make sense.