Add physical properties as first-class scene graph elements.
Velocity, mass, and material are added as first-class scene graph elements, to be sync’d client<->server. Mass = 0 is the default and this will indicate a static entity.

Currently material name is a string - it can be hardcoded to “default”, “ground”, or “vehicle”. In the future new scene tags will be added to make this more meaningful.

Building on this, the idea is that all properties of the physics engine are actually part of the scene graph, and so can be defined by world-builders and synchronised into clients.