Shift physics meta-data to become part of the scene.
By adding <physics>, <material>, and <contactmaterial> nodes, we can put all of the meta-data needed for the physics engine in as part of the scene. Right now, this means that scene designers can specify this data themselves.

In future, this will let us synchronise the physics engine with the client, but the client needs to be updated to support this.